Changed how Layout is used. renamed StaticQuery container to LayoutContainer to replicate redux naming
