+++ src/components/Layout/Layout.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { StaticQuery, graphql } from 'gatsby'
 import Helmet from 'react-helmet'
 import Header from '../Header'
 class Layout extends React.PureComponent {
@@ -31,44 +30,9 @@ class Layout extends React.PureComponent {
     }
 }
 
-class LayoutContainer extends React.PureComponent {
-    render() {
-        const { children } = this.props
-        return (
-            <StaticQuery
-                query={ graphql`
-                    query SiteTitleQuery {
-                        site {
-                            siteMetadata {
-                                title
-                                description
-                                author
-                            }
-                        }
-                    }
-                `}
-                render={ data => (
-                    <Layout
-                        siteMetadata={ data.site.siteMetadata }
-                        children={ children }
-                    />
-                )}
-            />
-        )
-    }
-}
-
 Layout.propTypes = {
     children: PropTypes.node.isRequired,
+    siteMetadata: PropTypes.object.isRequired
 }
 
-LayoutContainer.propTypes = {
-    children: PropTypes.node.isRequired,
-}
-
-// export default Layout
-
-export {
-    Layout,
-    LayoutContainer as default
-}
+export default Layout
