+++ src/components/Layout/Layout.stories.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
-import { Layout } from './Layout'
+import Layout from './Layout'
 
 storiesOf('Layout|Layout', module)
     .add('Basic', () => (
