+++ src/components/Layout/index.js
@@ -1,1 +1,2 @@
 export { default } from './Layout'
+export { default as LayoutContainer } from './LayoutContainer'
