+++ src/pages/404.js
@@ -1,12 +1,12 @@
 import React from 'react'
-import Layout from '../components/Layout'
+import { LayoutContainer } from '../components/Layout'
 
 const NotFoundPage = () => {
     return (
-        <Layout>
+        <LayoutContainer>
             <h1>NOT FOUND</h1>
             <p>You just hit a route that doesn&#39;t exist... the sadness.</p>
-        </Layout>
+        </LayoutContainer>
     )
 }
